backend/requirements-fastapi.txt
# FastAPI Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database (keeping SQLite - no additional packages needed)
# sqlite3 is built into Python

# Existing dependencies (if you have any)
# Add your existing requirements here
